Fix: 판매대금관리 레이아웃 변경
